fix(mcp): serve RFC 9728 oauth protected resource metadata at path-inserted well-known URL
